# Add faster `tanh` implementation(s)