--- a/src/activations.jl
+++ b/src/activations.jl
@@ -694,7 +694,7 @@ end
     exp2x = @fastmath exp(x + x)
     y = (exp2x - 1) / (exp2x + 1) 
     # That has large errors near zero; using `expm1` would more accurate, but about as slow as `tanh`.
-    # Instead, we switch to a polynomial, which is very accurate within its range, < 2 eps:
+    # Instead, we switch to a polynomial, which is very accurate within its range:
     x2 = x * x
     ypoly = x * evalpoly(x2, (1.0, -0.33333333333324583, 0.13333333325511604, -0.05396823125794372, 0.02186660872609521, -0.008697141630499953))
     ifelse(x2 > 900.0, sign(y), ifelse(x2 < 0.017, oftype(y, ypoly), y))