This proposes to add faster, lower-precision, versions of `tanh` and friends. Motivation:
```julia
julia> @btime x * x   setup=(x=randn(Float32,100,100));
  20.250 μs (2 allocations: 39.11 KiB)  # mac M1, julia 1.7 + rosetta
  52.718 μs (2 allocations: 39.11 KiB)  # xeon E5-2603

julia> @btime y .= tanh.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  107.167 μs (2 allocations: 39.11 KiB)
  229.725 μs (2 allocations: 39.11 KiB)

julia> @btime y .= tanh_faster.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  27.125 μs (2 allocations: 39.11 KiB)
  67.623 μs (2 allocations: 39.11 KiB)
```
There has been talk of using LoopVectorization to speed this up instead. At least for the computers I tried, this lower-precision function is faster. Of course that may speed up other things, but `tanh` seem the outlier in terms of speed -- I quote times with matrix multiplication above, to illustrate that it dominates. 
```julia
julia> using LoopVectorization

julia> @btime vmap!(tanh, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  62.542 μs (2 allocations: 39.11 KiB)
  99.213 μs (2 allocations: 39.11 KiB)

julia> @btime vmap!(tanh_faster, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));  # with IfElse
  27.292 μs (2 allocations: 39.11 KiB)
  60.970 μs (2 allocations: 39.11 KiB)
```
WIP, `tanh_fast` was the initial idea, but at least for Float32, a rational approximation like `tanh_faster` seems like a better idea, roughly one less digit & another factor of 3 in speed. Both need some care not to give crazy answers at large `x`, the precise implementation could use some adjusting. Times & error measures in [this gist](https://gist.github.com/mcabbott/0acd91671894bced65209e054bf80796).

I also have no idea hot this interacts with GPU just yet.