@DhairyaLGandhi 2021-08-07
We have been discussing a more structured approach to this lead by @chriselrod and NNlibCPU.