@chriselrod 2021-08-08
I could add `tanh_faster` [here](https://github.com/FluxML/NNlibCPU.jl/pull/1).
I wont be able to make the ML meeting this week, as I have a vet appointment at that time. But I'll try and merge that PR with basic dense, conv, and batch layers working.