@chriselrod 2021-08-08
Also, that Xeon in your benchmark seems to be very slow. Using:
```julia
using LoopVectorization, BenchmarkTools, LinearAlgebra
Sys.ARCH === :x86_64 && using MKL
BLAS.set_num_threads(1)
using IfElse: ifelse
@inline function tanh_faster(x::Real)
  T = float(typeof(x))
  x2 = 4x^2
  # y = 94.9339451088 * x * ((( x2 + 1.06315869e+03 ) * x2 + 1.88748783e+05 ) * x2 + 5.86237309e+06 ) / ((( ( x2 + 4.03183926e+03 ) * x2 + 1.64253046e+06 ) * x2 + 1.28592857e+08 ) * x2 + 1.11307745e+09 )
  num = @fastmath T(94.9339451088) * 2x * ((( x2 + T(1.06315869e+03) ) * x2 + T(1.88748783e+05) ) * x2 + T(5.86237309e+06) )
  den = @fastmath ((( ( x2 + T(4.03183926e+03) ) * x2 + T(1.64253046e+06) ) * x2 + T(1.28592857e+08) ) * x2 + T(1.11307745e+09) )
  y = num / den
  stop = _fast_stop(tanh_faster, x)
  ifelse(x > stop, one(y), ifelse(x < -stop, -one(y), y))
end
_fast_stop(::typeof(tanh_faster), x) = oftype(x, 10)

@btime x * x   setup=(x=randn(Float32,100,100));
@btime y .= tanh.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
@btime y .= tanh_faster.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
@btime vmap!(tanh, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
@btime vmap!(tanh_faster, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));  # with IfElse
```
I am getting:
```julia
julia> @btime x * x   setup=(x=randn(Float32,100,100));
  16.220 μs (2 allocations: 39.11 KiB) # i7-1165G7 (laptop)
  11.967 μs (2 allocations: 39.11 KiB) # i9-10980XE (HEDT)
  25.125 μs (2 allocations: 39.11 KiB) # M1 native

julia> @btime y .= tanh.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  75.780 μs (2 allocations: 39.11 KiB) # i7-1165G7 (laptop)
  94.014 μs (2 allocations: 39.11 KiB) # i9-10980XE (HEDT)
  83.917 μs (2 allocations: 39.11 KiB) # M1 native

julia> @btime y .= tanh_faster.(x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  19.565 μs (2 allocations: 39.11 KiB) # i7-1165G7 (laptop)
  15.447 μs (2 allocations: 39.11 KiB) # i9-10980XE (HEDT)
  28.417 μs (2 allocations: 39.11 KiB) # M1 native

julia> @btime vmap!(tanh, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));
  27.560 μs (2 allocations: 39.11 KiB) # i7-1165G7 (laptop)
  21.269 μs (2 allocations: 39.11 KiB) # i9-10980XE (HEDT)
  45.625 μs (2 allocations: 39.11 KiB) # M1 native

julia> @btime vmap!(tanh_faster, y, x * x)   setup=(x=randn(Float32,100,100); y=similar(x));  # with IfElse
  19.097 μs (2 allocations: 39.11 KiB) # i7-1165G7 (laptop)
  13.739 μs (2 allocations: 39.11 KiB) # i9-10980XE (HEDT)
  28.708 μs (2 allocations: 39.11 KiB) # M1 native
```