@mcabbott 2021-08-08
This was supposed to be a super-quick fix, but floating point stuff is messy!

Agree the nicest possible outcome would be a super-low-precision library. I think that for most ML purposes a smooth function which is <1pixel wrong on a graph is going to be plenty. My guess is that smoothness is more important than worst-case error, i.e. a steps are worse than ramps. NNlib might not be a terrible place for such things to live, though, if it's just `tanh` and some friends -- it's supposed to be a collection of NN-ish functions you can use for other purposes. They could move out once someone finds a need & time write other functions.

NNlibCPU.jl looks more ambitious. From very informal investigation there's lots of room for faster CPU `conv` etc.

If that's an optional add-on, then the argument for `tanh_faster` not being in there is that it might be useful without that. Motivated by things like [this thread](https://discourse.julialang.org/t/flux-vs-pytorch-cpu-performance/42667) -- it seems embarrassing that in the most obvious casual benchmark, 80% of the time is in `tanh`.