@oscardssmith 2021-08-08
One thing to try is using exp2 instead of exp. It may lose a tiny bit of precision, but be notably faster. exp needs some extra work to do the reduction, while exp2 gets it for free.