@ToucheSir 2021-08-08
Is PyTorch doing something extra besides using `tanh(f)`? If not, why would there still be a discrepancy between libm and https://github.com/JuliaLang/julia/pull/38382?