@mcabbott 2021-08-08
I never looked inside that PR somehow. The times for `tanh` will depend a lot on which branches you hit. And the time for `tanh_kernel` suggests that my rational `tanh_faster` has room for improvement:
```julia
julia> @btime y .= tanh.(x)  setup=(x = randn(Float32, 10^3); y = similar(x););  # as above
  2.963 μs (0 allocations: 0 bytes)

julia> @btime y .= tanh.(x)  setup=(x = 0.1f0 .* randn(Float32, 10^3); y = similar(x););  # small => kernel
  1.129 μs (0 allocations: 0 bytes)

julia> @btime y .= tanh.(x)  setup=(x = 100 .* randn(Float32, 10^3); y = similar(x););  # large => flat
  627.491 ns (0 allocations: 0 bytes)

julia> @btime y .= Base.Math.tanh_kernel.(x)  setup=(x = rand(Float32, 10^3); y = similar(x););
  144.404 ns (0 allocations: 0 bytes)

julia> Base.Math.TANH_SMALL_X(Float32)  # polynomial below this
1.3862944f0

julia> Base.Math.TANH_LARGE_X(Float32)  # flat above this, but why so large?
18.0f0

julia> tanh(9f0)  # already constant at half that.
1.0f0
```