@oscardssmith 2021-08-08
Just wondering, how did you generate the coefficients? Theoretically, Remez.jl can make minimax rational approximations, but I've never had much luck (I believe due to problems with numerical conditioning).