@mcabbott 2021-08-08
I copied them from here: https://math.stackexchange.com/questions/107292/rapid-approximation-of-tanhx