@mcabbott 2021-08-08
I never tried Remez but here's a quick attempt to find a function to cover the gap from the kernel to the flat (which presently calls `exp`). It indeed seems finicky as to what orders it can solve for, maybe there are tricks to make it behave better. Are there tricks for how to round the coefficients it finds to Float32?
```julia
using Remez, Plots
N,D,E,X = ratfn_minimax(tanh, (big"1.38", big"9.0"), 7, 4); Float32(E) # 1.2949664f-8
f(x) = evalpoly(x, N) / evalpoly(x, D)
xs = big.(1.3862944:0.01:9); plot(xs, Float64.(f.(xs) .- tanh.(xs))) # error in approx
E / eps(1f0) # 0.1
Tuple(Float32.(N))
Tuple(Float32.(D))
@inline function f74(x::Float32)
    n = evalpoly(x, (0.01731622f0, 0.9280657f0, 0.21158579f0, 0.08456047f0, 0.031536844f0, 0.00071064394f0, -3.1970143f-5, 6.226407f-7))
    d = evalpoly(x, (1.0f0, 0.08685698f0, 0.52843076f0, 0.016710332f0, 0.040428344f0))
    n / d
end
xs = 1.3862944:0.01:9; plot!(xs, (f74.(Float32.(xs)) .- tanh.(xs))) # error with Float32, up to 3e-7
3e-7 / eps(tanh(7.2f0)) # 5

using Base.Math: TANH_LARGE_X, TANH_SMALL_X, tanh_kernel
@inline function tanh_ifelse(x::T) where T<:Union{Float32}
    abs2x = abs(2x)
    ifelse(abs2x >= TANH_LARGE_X(T), copysign(one(T), x),
        ifelse(abs2x <= TANH_SMALL_X(T), x*tanh_kernel(x*x), 
            copysign(f74(abs(x)), x) ))  # clearly should re-use abs2x
end
xs = -10:0.01f0:10; plot(xs, tanh_ifelse.(xs)); plot!(xs, tanh.(xs))  # sanity check

@btime y .= tanh_ifelse.(x)  setup=(x = randn(Float32, 10^3); y = similar(x););
# 596.369 ns (0 allocations: 0 bytes)
# 3.109 μs (0 allocations: 0 bytes)  with branches instead of ifelse

# Accuracy of tanh_faster rational: 4e-7 until 8, and 1 - tanh(8f0) < 3e-7
xs = 0:0.001:8; plot(xs, @. Float64(abs(tanh_faster(Float32(xs)) - tanh(big(xs)))))
```

![tanh_errors](https://user-images.githubusercontent.com/32575566/128648552-925ba6d7-e5c4-4239-b3a6-7d4b059d2b6e.png)